requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
sentence-transformers
faiss-cpu
pypdf
openai
numpy
dotenv